+++ src/str.h
@@ -5,15 +5,15 @@
 
 typedef struct string{
     char* str;
-    size_t len;
-    size_t size;
+    int len;
+    int size;
 }string;
 
 int str_init(string *s);
 void str_reinit(string *s);
 void str_free(string *s);
 int str_add_char(string *s1, char c);
-int str_concat(string *s1, const char* s2, size_t len);
+int str_concat(string *s1, const char* s2, int len);
 int str_copy(string *s1, const string *s2);
 int str_cmp(const string *s1, const string *s2);
 int str_cmp_c_str(const string *s1, const char* s2);
